refactor: Remove get_reg out of rmi interface
Previously, `get_reg` was used to sidestep the dependency issue between
monitor and armv9a. Now that the dependency issue has been resolved, it
doesn't have to be inside the list of RMI's Interface.
