+++ rmm/src/rmi/rec/exit.rs
@@ -1,6 +1,7 @@
 use crate::event::realmexit::*;
 use crate::event::{Context, RsiHandle};
 use crate::granule::GRANULE_MASK;
+use crate::realm::context::get_reg;
 use crate::realm::mm::stage2_tte::S2TTE;
 use crate::rmi::error::Error;
 use crate::rmi::rec::run::Run;
@@ -83,12 +84,12 @@ fn is_non_emulatable_data_abort(
     Ok(ret)
 }
 
-fn get_write_val(rmi: RMI, realm_id: usize, vcpu_id: usize, esr_el2: u64) -> Result<u64, Error> {
+fn get_write_val(_rmi: RMI, realm_id: usize, vcpu_id: usize, esr_el2: u64) -> Result<u64, Error> {
     let esr_el2 = EsrEl2::new(esr_el2);
     let rt = esr_el2.get_masked_value(EsrEl2::SRT) as usize;
     let write_val = match rt == 31 {
         true => 0, // xzr
-        false => rmi.get_reg(realm_id, vcpu_id, rt)? as u64 & esr_el2.get_access_size_mask(),
+        false => get_reg(realm_id, vcpu_id, rt)? as u64 & esr_el2.get_access_size_mask(),
     };
     Ok(write_val)
 }
