+++ rmm/src/rsi/mod.rs
@@ -11,7 +11,7 @@ use crate::listen;
 use crate::measurement::{
     HashContext, Measurement, MeasurementError, MEASUREMENTS_SLOT_NR, MEASUREMENTS_SLOT_RIM,
 };
-use crate::realm::context::set_reg;
+use crate::realm::context::{get_reg, set_reg};
 use crate::realm::mm::address::GuestPhysAddr;
 use crate::realm::mm::stage2_tte::invalid_ripas;
 use crate::rmi;
@@ -49,15 +49,14 @@ extern crate alloc;
 pub fn do_host_call(
     _arg: &[usize],
     ret: &mut [usize],
-    rmm: &Monitor,
+    _rmm: &Monitor,
     rec: &mut Rec<'_>,
     run: &mut Run,
 ) -> core::result::Result<(), Error> {
-    let rmi = rmm.rmi;
     let vcpuid = rec.vcpuid();
     let realmid = rec.realmid()?;
 
-    let ipa = rmi.get_reg(realmid, vcpuid, 1).unwrap_or(0x0);
+    let ipa = get_reg(realmid, vcpuid, 1).unwrap_or(0x0);
 
     let pa = crate::realm::registry::get_realm(realmid)
         .ok_or(Error::RmiErrorOthers(NotExistRealm))?
@@ -113,15 +112,14 @@ pub trait Interface {
 }
 
 pub fn set_event_handler(rsi: &mut RsiHandle) {
-    listen!(rsi, ATTEST_TOKEN_INIT, |_arg, ret, rmm, rec, _| {
-        let rmi = rmm.rmi;
+    listen!(rsi, ATTEST_TOKEN_INIT, |_arg, ret, _rmm, rec, _| {
         let realmid = rec.realmid()?;
         let vcpuid = rec.vcpuid();
 
         let mut challenge: [u8; 64] = [0; 64];
 
         for i in 0..8 {
-            let challenge_part = rmi.get_reg(realmid, vcpuid, i + 2)?;
+            let challenge_part = get_reg(realmid, vcpuid, i + 2)?;
             let start_idx = i * 8;
             let end_idx = start_idx + 8;
             challenge[start_idx..end_idx].copy_from_slice(&challenge_part.to_le_bytes());
@@ -140,7 +138,6 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
     });
 
     listen!(rsi, ATTEST_TOKEN_CONTINUE, |_arg, ret, rmm, rec, _| {
-        let rmi = rmm.rmi;
         let realmid = rec.realmid()?;
         let ipa_bits = rec.ipa_bits()?;
 
@@ -157,7 +154,7 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
             return Ok(());
         }
 
-        let attest_ipa = rmi.get_reg(realmid, vcpuid, 1)?;
+        let attest_ipa = get_reg(realmid, vcpuid, 1)?;
         if validate_ipa(attest_ipa, ipa_bits).is_err() {
             warn!("Wrong ipa passed {}", attest_ipa);
             set_reg(realmid, vcpuid, 0, ERROR_INPUT)?;
@@ -210,11 +207,10 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
     });
 
     listen!(rsi, MEASUREMENT_READ, |_arg, ret, rmm, rec, _| {
-        let rmi = rmm.rmi;
         let vcpuid = rec.vcpuid();
         let realmid = rec.realmid()?;
         let mut measurement = Measurement::empty();
-        let index = rmi.get_reg(realmid, vcpuid, 1)?;
+        let index = get_reg(realmid, vcpuid, 1)?;
 
         if index >= MEASUREMENTS_SLOT_NR {
             warn!("Wrong index passed: {}", index);
@@ -239,20 +235,16 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
     });
 
     listen!(rsi, MEASUREMENT_EXTEND, |_arg, ret, rmm, rec, _| {
-        let rmi = rmm.rmi;
         let vcpuid = rec.vcpuid();
         let realmid = rec.realmid()?;
 
-        let index = rmi.get_reg(realmid, vcpuid, 1)?;
-        let size = rmi.get_reg(realmid, vcpuid, 2)?;
+        let index = get_reg(realmid, vcpuid, 1)?;
+        let size = get_reg(realmid, vcpuid, 2)?;
         let mut buffer = [0u8; 64];
 
         for i in 0..8 {
-            buffer[i * 8..i * 8 + 8].copy_from_slice(
-                rmi.get_reg(realmid, vcpuid, i + 3)?
-                    .to_le_bytes()
-                    .as_slice(),
-            );
+            buffer[i * 8..i * 8 + 8]
+                .copy_from_slice(get_reg(realmid, vcpuid, i + 3)?.to_le_bytes().as_slice());
         }
 
         if size > buffer.len() || index == MEASUREMENTS_SLOT_RIM || index >= MEASUREMENTS_SLOT_NR {
@@ -279,7 +271,7 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
         let vcpuid = rec.vcpuid();
         let ipa_bits = rec.ipa_bits()?;
         let realmid = rec.realmid()?;
-        let config_ipa = rmi.get_reg(realmid, vcpuid, 1)?;
+        let config_ipa = get_reg(realmid, vcpuid, 1)?;
         if validate_ipa(config_ipa, ipa_bits).is_err() {
             set_reg(realmid, vcpuid, 0, ERROR_INPUT)?;
             ret[0] = rmi::SUCCESS_REC_ENTER;
@@ -304,7 +296,7 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
         let ipa_bits = rec.ipa_bits()?;
         let realmid = rec.realmid()?;
 
-        let ipa_page = rmi.get_reg(realmid, vcpuid, 1)?;
+        let ipa_page = get_reg(realmid, vcpuid, 1)?;
         if validate_ipa(ipa_page, ipa_bits).is_err() {
             if set_reg(realmid, vcpuid, 0, ERROR_INPUT).is_err() {
                 warn!(
@@ -341,15 +333,14 @@ pub fn set_event_handler(rsi: &mut RsiHandle) {
         Ok(())
     });
 
-    listen!(rsi, IPA_STATE_SET, |_arg, ret, rmm, rec, run| {
-        let rmi = rmm.rmi;
+    listen!(rsi, IPA_STATE_SET, |_arg, ret, _rmm, rec, run| {
         let vcpuid = rec.vcpuid();
         let realmid = rec.realmid()?;
         let ipa_bits = rec.ipa_bits()?;
 
-        let ipa_start = rmi.get_reg(realmid, vcpuid, 1)?;
-        let ipa_size = rmi.get_reg(realmid, vcpuid, 2)?;
-        let ipa_state = rmi.get_reg(realmid, vcpuid, 3)? as u8;
+        let ipa_start = get_reg(realmid, vcpuid, 1)?;
+        let ipa_size = get_reg(realmid, vcpuid, 2)?;
+        let ipa_state = get_reg(realmid, vcpuid, 3)? as u8;
         let ipa_end = ipa_start + ipa_size;
 
         if ipa_end <= ipa_start {
